refactor: modularize card handler mixins
